Heavy Refactor Part 1, move round clues back up to the top.
